backend/pyproject.toml
[tool.poetry]
name = "f1-pivot"
version = "1.0"
description = ""
authors = ["Bradley Harmeyer <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = { extras = ["all"], version = ">=0.115.12,<0.116.0" }
uvicorn = { extras = ["standard"], version = ">=0.34.2,<0.35.0" }
sqlalchemy = ">=2.0.40,<3.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
requests = "^2.32.4"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
